requirements.txt
anthropic
openai
python-dotenv
sentence-transformers
chromadb
requests
pytest
